+++ tests/test_components.py
@@ -93,6 +93,32 @@ def test_entity_projectile_on_hit_methods():
     }
 
     proj = EntityProjectile()
+    proj.impact_damage()
+    assert proj._component["on_hit"]["impact_damage"] == {}
+
+    proj = EntityProjectile()
+    proj.impact_damage(damage=4, filter="zombie", difficulty_randomization="additive")
+    assert proj._component["on_hit"]["impact_damage"] == {
+        "damage": {"min": 4, "max": 4},
+        "filter": "zombie",
+        "difficulty_randomization": "additive",
+    }
+
+    proj = EntityProjectile()
+    proj.impact_damage(damage={"max": 6, "min": 3}, power_multiplier=2.0)
+    assert proj._component["on_hit"]["impact_damage"] == {
+        "damage": {"min": 3, "max": 6},
+        "power_multiplier": 2.0,
+    }
+
+    with pytest.raises(ValueError):
+        EntityProjectile().impact_damage(difficulty_randomization="random")
+    with pytest.raises(ValueError):
+        EntityProjectile().impact_damage(damage={"minimum": 1})
+    with pytest.raises(ValueError):
+        EntityProjectile().impact_damage(min_critical_damage=10, max_critical_damage=5)
+
+    proj = EntityProjectile()
     proj.spawn_chance(
         spawn_definition="minecraft:chicken",
         first_spawn_chance=0.5,